Entities in f64
The line stage 42 drew at the things that travel, moved out to every body
in the game. Villagers, deputies, garrison holders, the stalker, shots,
sweeps and impacts, falling stems, kestrel marks, sightings, the belief
board's last-known, caravans and crashes carry DVec3 positions; directions,
headings, ranges, speeds and rig geometry stay f32, and the difference is
taken wide before it is narrowed. Every .as_vec3() that handed the player
to an entity system and every .as_dvec3() that widened the entity back at
a sight call is gone.

The draw path builds every body in the camera's frame: Rig::objects takes
the local Vec3 from renderer.relative(position) and the objects are marked
already_relative, as the held tool and the player's body were; Villagers,
Mining and kestrel_objects take the relative closure; the fixtures place
the camera first and build through Camera::relative. The per-object
lighting pass adds the render origin back before reading a column, which
had lit the player's own body from the wrong chunk since stage 42. Two
duplicated draw blocks (roost kestrel, marks) collapsed to one.

The muzzle of a shot crosses the journal's wire at f64 (VERSION 29; a
version-28 log restarts the oracle as every older log does) and starts at
the eye. arsenal.dat is VERSION 2 with f64 crash columns and a v1 reader
that widens the f32 ones.

Proofs, in the stage-42 shape: the same posse walks the same walk, the
same townsfolk stroll the same stroll, the same rounds leave the same
craters block for block and mask for mask, and the same rig draws the same
bytes, at spawn and three thousand kilometres out; a hand-written v1
arsenal.dat widens and a v2 far column round-trips.

On the way past: a stem landing across a felled tree's stump felled it
again, and the chain count was read off the landing tick alone, both
exposed by the level tip no longer sitting a millionth under the hinge;
and the handheld's roster listed the kestrel twice.

+++ crates/vx-app/src/economy.rs
@@ -477,13 +477,13 @@ impl Shipment {
 
     /// Where the load is at `now`, as a column. Pure in the clock: same tick,
     /// same place, however often it is asked.
-    pub fn position_at(&self, now: u64) -> (f32, f32) {
-        let span = self.arrive.saturating_sub(self.depart).max(1) as f32;
-        let travelled = now.saturating_sub(self.depart) as f32;
+    pub fn position_at(&self, now: u64) -> (f64, f64) {
+        let span = self.arrive.saturating_sub(self.depart).max(1) as f64;
+        let travelled = now.saturating_sub(self.depart) as f64;
         let t = (travelled / span).clamp(0.0, 1.0);
         (
-            self.from.0 as f32 + (self.to.0 - self.from.0) as f32 * t,
-            self.from.1 as f32 + (self.to.1 - self.from.1) as f32 * t,
+            f64::from(self.from.0) + f64::from(self.to.0 - self.from.0) * t,
+            f64::from(self.from.1) + f64::from(self.to.1 - self.from.1) * t,
         )
     }
 }
